[FEATURE] Move and rename Extractor to ImageDimensionsExtraction
Move Extractor to Service/Extraction namespace and rename Extractor
to ImageDimensionsExtraction
